New Grid Layout
Changed the shelves to be setup using a Grid layout rather than flexing across in a single row. This allows for more verticality and better accuracy for how far left or right an item is. The parts .CSV file (and translated .js file) now have four extra properties: PosX, PosY, SizeX, and SizeY to describe the position of the bin on the shelf. There was some refactoring of the toggleItemDisplay() method to use a more succinct and easier to read tool (classList.toggle()) and a nicer looking transition for when a bin is selected and deselected.