Improved Searching
Searching is a lot better now. The search bar is now a form, so pressing Enter will submit the search data and no longer requires the user to press the search button. If a search doesn't have anything to show, it will display an alert that tells the user it couldn't find what they searched for. When searching, it will smoothly scroll to the first instance of the search. The header, which contains the search bar, has been made sticky so that it follows as the page scrolls, allowing for easier navigation to add more searches.